Modify VariableClient (TF2) asynchronous update logic.
This is done as a safety measure to avoid silent variable corruption in the
middle of a `select_action()` call.

Note that this means that variables can only be copied to the TF graph on a
subsequent call to `update()`. This is not usually a problem as the latter is
usually called within a high-frequency (relative to the update_period) loop;
but there may be edge cases. Consider using `VariableClient.update_and_wait()`
inside your actor's update method if this becomes an issue.

PiperOrigin-RevId: 319044266
Change-Id: I519188bd5f5221614f3939ca5c062a3a80cbdee5