Add a Builder class which encapsulates a full agent.
This class allows for the consituent components to be broken apart so that it
can be used both for distributed and non-distributed variants. For the
time-being this is only incorporated into the TF D4PG agent to allow for
minimal disruption and experimentation, but should be rolled out for all agents
soon.

PiperOrigin-RevId: 356975846
Change-Id: I00ead33da40f4f98052ae3beb218c23788ada206