SVG-0 agent with entropy and behavior prior.
The stochastic value gradients (SVG)-0 agent bears similarity to DPG but uses the reparameterization trick to learn stochastic policies. This version uses a continuous version of RETRACE to learn the value function across multiple steps and an added entropy bonus for the actor loss. Additionally the code supports learning behavior priors with an example on running with the locomotion go to target task.

PiperOrigin-RevId: 390159206
Change-Id: Ide9d8fd3dca9d6513ac571becd748a701e3c0173