Speed up get_from_first_device.
If the nested structure has many large fields then it is important that we trigger the D2H transfer for all leaves and then start converting to numpy. As it is now the copy of the next leaf will not start until the previous one has completed its conversion to a numpy array (which requires a copy of the array content).

PiperOrigin-RevId: 395886644
Change-Id: I0671935932a619d48b223acd023852a5dc8992f2